+++ src/GitVersion.Core/VersionCalculation/SemanticVersioning/SemanticVersionPreReleaseTag.cs
@@ -15,7 +15,7 @@ public class SemanticVersionPreReleaseTag :
     {
     }
 
-    public SemanticVersionPreReleaseTag(string? name, int? number)
+    public SemanticVersionPreReleaseTag(string? name, long? number)
     {
         Name = name;
         Number = number;
@@ -29,7 +29,7 @@ public class SemanticVersionPreReleaseTag :
     }
 
     public string? Name { get; set; }
-    public int? Number { get; set; }
+    public long? Number { get; set; }
     public bool? PromotedFromCommits { get; set; }
 
     public override bool Equals(object? obj) => Equals(obj as SemanticVersionPreReleaseTag);
@@ -76,7 +76,7 @@ public class SemanticVersionPreReleaseTag :
         }
 
         var value = match.Groups["name"].Value;
-        var number = match.Groups["number"].Success ? int.Parse(match.Groups["number"].Value) : (int?)null;
+        var number = match.Groups["number"].Success ? long.Parse(match.Groups["number"].Value) : (long?)null;
         if (value.EndsWith("-"))
             return new SemanticVersionPreReleaseTag(preReleaseTag, null);
